requirements.txt
fastapi
uvicorn
numpy
tensorflow
firebase-admin
Pillow
opencv-python
python-multipart
python-dotenv
